prep tests for benchctx
