+++ core/integration/module_up_test.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
+	"testing"
 	"time"
 
 	"github.com/creack/pty"
@@ -16,8 +17,11 @@ import (
 
 // Note: ensure each testcase use unique port, otherwise you might see flakes.
 func (ModuleSuite) TestDaggerUp(ctx context.Context, t *testctx.T) {
+	if testing.Short() {
+		t.SkipNow()
+	}
 	// set timeout to 3m for each test
-	t = t.WithTimeout(3 * time.Minute)
+	t = t.WithTimeout(3 * time.Minute).(*testctx.T)
 
 	const defaultTrafficPortForContainerTests = "23100"
 	const defaultTrafficPortForServiceTests = "23200"
@@ -99,7 +103,7 @@ func (ModuleSuite) TestDaggerUp(ctx context.Context, t *testctx.T) {
 		},
 	}
 
-	t = t.WithTimeout(3 * time.Minute)
+	t = t.WithTimeout(3 * time.Minute).(*testctx.T)
 
 	for _, tc := range testcases {
 		t.Run(tc.name, func(ctx context.Context, t *testctx.T) {
@@ -183,7 +187,7 @@ func daggerUpInitModFn(ctx context.Context, t *testctx.T, defaultPort string) st
 		"strconv"
 		"dagger/test/internal/dagger"
 	)
-	
+
 	func New(
 		// +optional
 		// +default=%s
@@ -201,7 +205,7 @@ func daggerUpInitModFn(ctx context.Context, t *testctx.T, defaultPort string) st
 				WithDefaultArgs([]string{"python", "-m", "http.server", strconv.Itoa(port)}),
 		}
 	}
-	
+
 	type Test struct {
 		Ctr *dagger.Container
 	}
